+++ bin/node-credentials.js
@@ -1,82 +1,6 @@
 #! /usr/bin/env node
-const fs = require("fs");
+const cli = require("../src/cli");
 const commandLineArgs = require("command-line-args");
-const commandLineUsage = require("command-line-usage");
-const Vault = require("../src/vault").Vault;
-
-const encrypt = ({ path }) => {
-  const vault = new Vault({
-    credentialsFilePath: `${path}/credentials.json`
-  });
-  const vault = new Vault();
-  if (fs.existsSync(`${vault.credentialsFilePath}.key`)) {
-    vault.encryptFile();
-    fs.unlinkSync(`${vault.credentialsFilePath}`);
-  } else {
-    console.log(
-      "Warning credentials.json.key not exists, create new key with init"
-    );
-  }
-};
-
-const edit = ({ path }) => {
-  const vault = new Vault({
-    credentialsFilePath: `${path}/credentials.json`
-  });
-  const vault = new Vault();
-  if (!fs.existsSync(`${vault.credentialsFilePath}.enc`)) {
-    console.log("Error credentials.json.enc not exists");
-  } else {
-    vault.editCredentials();
-    fs.unlinkSync(`${vault.credentialsFilePath}.enc`);
-  }
-};
-
-const init = ({ path }) => {
-  const vault = new Vault({ credentialsFilePath: `${path}/credentials.json` });
-  if (fs.existsSync(`${vault.credentialsFilePath}.enc`)) {
-    console.log(
-      "Warning credentials.json.enc exists, ensure decrypt file before generate new key"
-    );
-  } else if (fs.existsSync(`${vault.credentialsFilePath}.key`)) {
-    console.log(
-      "Warning credentials.json.key exists, delete credentials.json.key to generate new key"
-    );
-  } else {
-    vault.createNewKey();
-    vault.encryptFile();
-    fs.unlinkSync(`${vault.credentialsFilePath}`);
-  }
-};
-
-const help = () => {
-  const sections = [
-    {
-      header: "node-vault",
-      content: "encrypted your credentials"
-    },
-    {
-      header: "Synopsis",
-      content: "node-vault <command> <options>"
-    },
-    {
-      header: "Command List",
-      content: [
-        { name: "help", summary: "help" },
-        {
-          name: "init",
-          summary:
-            "create credentials.json.key and encrypt your credentials.json"
-        },
-        { name: "encrypt", summary: "encrypt credentials.json" },
-        { name: "decrypt", summary: "decrypt credentials.json.enc" },
-        { name: "edit", summary: "decrypt alias" }
-      ]
-    }
-  ];
-  const usage = commandLineUsage(sections);
-  console.log(usage);
-};
 
 const parseCommand = () => {
   const mainDefinitions = [{ name: "command", defaultOption: true }];
@@ -94,7 +18,6 @@ const parseCommand = () => {
 
 const { command, options } = parseCommand();
 const uknowFnc = () => console.log("invalid command");
-const commandCase = { encrypt, edit, decrypt: edit, init, help };
 
-const commandFnc = commandCase[command] || uknowFnc;
+const commandFnc = cli[command] || uknowFnc;
 commandFnc(options);
